requirements.txt
mysql_connector_repackaged==0.3.1
python-dotenv==0.15.0

mysql-connector-python~=8.0.22
Flask~=1.1.2